FEAT: created Junit test implemenation, but test will not be run unti…

+++ Anthony_John_Order_Parts_CaseStudy/src/main/java/com/example/Anthony_John_Order_Parts_CaseStudy/entity/AircraftPartInventoryEntity.java
@@ -25,7 +25,7 @@ public class AircraftPartInventoryEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column
-    private Long id;  // PK for the aircraft part
+    private int id;  // PK for the aircraft part
 
     // IMPORT THE AIRCRAFT PART MODEL
     @ManyToOne
